refactor(runtime): extract model input and run events
