+++ .changeset/clean-runtime-boundaries.md
@@ -1,0 +1,5 @@
+---
+"@opengeni/runtime": patch
+---
+
+Split model-input shaping and SDK run-event normalization into package-private runtime modules while preserving the existing public entrypoint and exports.
